+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2024-10-04
+
+## Added
+
+- Added internal, Speechmatics only client message: GetSpeakers, and server message: SpeakersResult
+- Added internal, Speechmatics only client method: send_message
+
 ## [2.0.1] - 2024-08-28
 
 ### Changed